+++ test/libsolidity/syntaxTests/indexing/index_range_access_assert.sol
@@ -3,4 +3,3 @@ contract s{}
 function f() {s[:][];}
 // ----
 // TypeError 1760: (53-57): Types cannot be sliced.
-// TypeError 2876: (53-59): Index access for contracts or libraries is not possible.